requirements.txt
# --- Frontend & UI ---
streamlit>=1.30.0

# --- Database Connectivity ---
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0

# --- Cloud Storage (Azure) ---
azure-storage-blob>=12.19.0

# --- Machine Learning & Data Processing ---
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.4.2
imbalanced-learn==0.12.3
joblib>=1.3.0

# --- Utilities ---
requests>=2.31.0

# --- Analytics (Optional for evaluation) ---
matplotlib>=3.7.0

python-dotenv>=1.0.0